split BarChar and StackedBarChart classes
