+++ src/sentry/static/sentry/app/views/projectDashboard/chart.jsx
@@ -82,7 +82,7 @@ const ProjectChart = React.createClass({
 
   renderChart() {
     let points = this.state.stats.map((point) => {
-      return {x: point[0], y: [point[1]]};
+      return {x: point[0], y: point[1]};
     });
     let startX = (new Date().getTime() / 1000) - 3600 * 24 * 7;
     let markers = this.state.releaseList.filter((release) => {
